Write stubs for chunkMeshing system
About the design: each chunk can have as many "mesh layers" as it wants but there can only be one "face set" and one "mesh instance" for each face inside the chunk. The VoxelChunkMesher is in no way responsible for updating buffers, encoding vertices, forming faces, or interacting with the GPU. All voxels are programatically described using a "voxel_describer" function which specifies which faces and mesh instance a given voxel will have, and is generally in charge of managing mesh layers.